Extract GRPC Test Service as package
